Correction type de montant et tva dans entity + adaptation CrudController for MoneyField + PercentField

+++ src/Entity/Article.php
@@ -76,7 +76,7 @@ class Article
     private $detailCommandeArticles;
 
     /**
-     * @ORM\Column(type="float")  
+     * @ORM\Column(type="float", scale=2)  
      */
     private $tauxTva;
 
@@ -130,12 +130,12 @@ class Article
         return $this->id;
     }
 
-    public function getMontantHorsTva(): ?int
+    public function getMontantHorsTva(): ?float
     {
         return $this->montantHorsTva;
     }
 
-    public function setMontantHorsTva(int $montantHorsTva): self
+    public function setMontantHorsTva(float $montantHorsTva): self
     {
         $this->montantHorsTva = $montantHorsTva;
 
@@ -367,12 +367,12 @@ class Article
         return $this;
     }
 
-    public function getTauxTva(): ?int
+    public function getTauxTva(): ?float
     {
         return $this->tauxTva;
     }
 
-    public function setTauxTva(int $tauxTva): self
+    public function setTauxTva(float $tauxTva): self
     {
         $this->tauxTva = $tauxTva;
