+++ src/Entity/Reservation.php
@@ -35,17 +35,17 @@ class Reservation
     private $facture;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float", scale=2)
      */
     private $montantTotalHorsTva;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float", scale=2)
      */
     private $montantTva;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float", scale=2)
      */
     private $montantTotal;
 
@@ -90,36 +90,36 @@ class Reservation
         return $this;
     }
 
-    public function getMontantTotalHorsTva(): ?int
+    public function getMontantTotalHorsTva(): ?float
     {
         return $this->montantTotalHorsTva;
     }
 
-    public function setMontantTotalHorsTva(int $montantTotalHorsTva): self
+    public function setMontantTotalHorsTva(float $montantTotalHorsTva): self
     {
         $this->montantTotalHorsTva = $montantTotalHorsTva;
 
         return $this;
     }
 
-    public function getMontantTva(): ?int
+    public function getMontantTva(): ?float
     {
         return $this->montantTva;
     }
 
-    public function setMontantTva(int $montantTva): self
+    public function setMontantTva(float $montantTva): self
     {
         $this->montantTva = $montantTva;
 
         return $this;
     }
 
-    public function getMontantTotal(): ?int
+    public function getMontantTotal(): ?float
     {
         return $this->montantTotal;
     }
 
-    public function setMontantTotal(int $montantTotal): self
+    public function setMontantTotal(float $montantTotal): self
     {
         $this->montantTotal = $montantTotal;
